+++ lib/busybee/testing/helpers.rb
@@ -1,9 +1,8 @@
 # frozen_string_literal: true
 
-require "active_support"
-require "active_support/duration"
 require "base64"
 require "securerandom"
+require "busybee/durations"
 require "busybee/grpc"
 require "busybee/serialization"
 require "busybee/testing/activated_job"
@@ -184,7 +183,7 @@ module Busybee
       # @param vars [Hash] message variables
       # @param ttl [Integer, ActiveSupport::Duration] time-to-live (integer ms or Duration)
       def publish_message(name, correlation_key:, vars: {}, ttl: 5000)
-        ttl_ms = ttl.is_a?(ActiveSupport::Duration) ? ttl.in_milliseconds.to_i : ttl.to_i
+        ttl_ms = Busybee::Durations.milliseconds_from(ttl)
         request = Busybee::GRPC::PublishMessageRequest.new(
           name: name,
           correlationKey: correlation_key,