+++ lib/busybee/testing/helpers/support.rb
@@ -1,9 +1,8 @@
 # frozen_string_literal: true
 
-require "active_support"
-require "active_support/duration"
 require "json"
 require "securerandom"
+require "busybee/durations"
 require "busybee/grpc"
 
 module Busybee
@@ -44,11 +43,7 @@ module Busybee
           worker = "#{type}-#{SecureRandom.hex(4)}"
 
           request_timeout = timeout || Busybee.default_job_request_timeout
-          request_timeout_ms = if request_timeout.is_a?(ActiveSupport::Duration)
-                                 request_timeout.in_milliseconds.to_i
-                               else
-                                 request_timeout.to_i
-                               end
+          request_timeout_ms = Busybee::Durations.milliseconds_from(request_timeout)
 
           request = Busybee::GRPC::ActivateJobsRequest.new(
             type: type,